+++ src/main/java/com/private_void/core/geometry/Point3D.java
@@ -1,7 +1,7 @@
 package com.private_void.core.geometry;
 
 public interface Point3D {
-    float getQ1();
-    float getQ2();
-    float getQ3();
+    double getQ1();
+    double getQ2();
+    double getQ3();
 }
