+++ src/main/java/com/private_void/core/geometry/Vector.java
@@ -3,20 +3,20 @@ package com.private_void.core.geometry;
 // TODO: Положительным углам при этом соответствует вращение вектора против часовой стрелки в правой системе координат, и по часовой стрелке в левой системе координат, если смотреть против направления соответствующей оси
 
 public class Vector extends CartesianPoint {
-    private float[][] rotationMatrixXYZ;
-    private float[][] rotationMatrixX;
-    private float[][] rotationMatrixY;
+    private double[][] rotationMatrixXYZ;
+    private double[][] rotationMatrixX;
+    private double[][] rotationMatrixY;
 
-    public Vector(float x, float y, float z) {
+    public Vector(double x, double y, double z) {
         super(x, y, z);
-        rotationMatrixXYZ = new float[3][3];
-        rotationMatrixX = new float[3][3];
-        rotationMatrixY = new float[3][3];
+        rotationMatrixXYZ = new double[3][3];
+        rotationMatrixX = new double[3][3];
+        rotationMatrixY = new double[3][3];
         normalize();
     }
 
     private void normalize() {
-        float norm = (float) Math.sqrt(x * x + y * y + z * z);
+        double norm =  Math.sqrt(x * x + y * y + z * z);
         try {
             x = x / norm;
             y = y / norm;
@@ -34,8 +34,8 @@ public class Vector extends CartesianPoint {
         return this;
     }
 
-    public Vector turnAroundOX(float angle) {
-        float[] temp = {0.0f, 0.0f, 0.0f};
+    public Vector turnAroundOX(double angle) {
+        double[] temp = {0.0, 0.0, 0.0};
         setRotationMatrixX(angle);
 
         for (int i = 0; i < 3; i++) {
@@ -50,8 +50,8 @@ public class Vector extends CartesianPoint {
         return this;
     }
 
-    public Vector turnAroundOY(float angle) {
-        float[] temp = {0.0f, 0.0f, 0.0f};
+    public Vector turnAroundOY(double angle) {
+        double[] temp = {0.0, 0.0, 0.0};
         setRotationMatrixY(angle);
 
         for (int i = 0; i < 3; i++) {
@@ -66,8 +66,8 @@ public class Vector extends CartesianPoint {
         return this;
     }
 
-    public Vector turnAroundVector(float angle, final Vector axis) {
-        float[] temp = {0.0f, 0.0f, 0.0f};
+    public Vector turnAroundVector(double angle, final Vector axis) {
+        double[] temp = {0.0, 0.0, 0.0};
         setRotationMatrixXYZ(angle, axis);
 
         for (int i = 0; i < 3; i++) {
@@ -82,8 +82,8 @@ public class Vector extends CartesianPoint {
         return this;
     }
 
-    public Vector getNewByTurningAroundOX(float angle) {
-        float[] temp = {0.0f, 0.0f, 0.0f};
+    public Vector getNewByTurningAroundOX(double angle) {
+        double[] temp = {0.0, 0.0, 0.0};
         setRotationMatrixX(angle);
 
         for (int i = 0; i < 3; i++) {
@@ -95,8 +95,8 @@ public class Vector extends CartesianPoint {
         return result;
     }
 
-    public Vector getNewByTurningAroundVector(float angle, final Vector axis) {
-        float[] temp = {0.0f, 0.0f, 0.0f};
+    public Vector getNewByTurningAroundVector(double angle, final Vector axis) {
+        double[] temp = {0.0, 0.0, 0.0};
         setRotationMatrixXYZ(angle, axis);
 
         for (int i = 0; i < 3; i++) {
@@ -108,67 +108,68 @@ public class Vector extends CartesianPoint {
         return result;
     }
 
-    public float getAngle(final Vector vector) {
-        return (float) (Math.acos((x * vector.getX() + y * vector.getY() + z * vector.getZ()) /
-                       (Math.sqrt(x * x + y * y + z * z) * Math.sqrt(vector.getX() * vector.getX() + vector.getY() * vector.getY() + vector.getZ() * vector.getZ()))));
+    public double getAngle(final Vector vector) {
+        return Math.acos((x * vector.getX() + y * vector.getY() + z * vector.getZ()) 
+                / (Math.sqrt(x * x + y * y + z * z) * Math.sqrt(vector.getX() * vector.getX() 
+                + vector.getY() * vector.getY() + vector.getZ() * vector.getZ())));
     }
 
-    private void setRotationMatrixX(float angle) {
-        rotationMatrixX[0][0] = 1.0f;
-        rotationMatrixX[1][0] = 0.0f;
-        rotationMatrixX[2][0] = 0.0f;
+    private void setRotationMatrixX(double angle) {
+        rotationMatrixX[0][0] = 1.0;
+        rotationMatrixX[1][0] = 0.0;
+        rotationMatrixX[2][0] = 0.0;
 
-        rotationMatrixX[0][1] = 0.0f;
-        rotationMatrixX[1][1] = (float) Math.cos(angle);
-        rotationMatrixX[2][1] = (float) -Math.sin(angle);
+        rotationMatrixX[0][1] = 0.0;
+        rotationMatrixX[1][1] = Math.cos(angle);
+        rotationMatrixX[2][1] = -Math.sin(angle);
 
-        rotationMatrixX[0][2] = 0.0f;
-        rotationMatrixX[1][2] = (float) Math.sin(angle);
-        rotationMatrixX[2][2] = (float) Math.cos(angle);
+        rotationMatrixX[0][2] = 0.0;
+        rotationMatrixX[1][2] = Math.sin(angle);
+        rotationMatrixX[2][2] = Math.cos(angle);
     }
 
-    private void setRotationMatrixY(float angle) {
-        rotationMatrixY[0][0] = (float) Math.cos(angle);
-        rotationMatrixY[1][0] = 0.0f;
-        rotationMatrixY[2][0] = (float) Math.sin(angle);
+    private void setRotationMatrixY(double angle) {
+        rotationMatrixY[0][0] = Math.cos(angle);
+        rotationMatrixY[1][0] = 0.0;
+        rotationMatrixY[2][0] = Math.sin(angle);
 
-        rotationMatrixY[0][1] = 0.0f;
-        rotationMatrixY[1][1] = 1.0f;
-        rotationMatrixY[2][1] = 0.0f;
+        rotationMatrixY[0][1] = 0.0;
+        rotationMatrixY[1][1] = 1.0;
+        rotationMatrixY[2][1] = 0.0;
 
-        rotationMatrixY[0][2] = (float) -Math.sin(angle);
-        rotationMatrixY[1][2] = 0.0f;
-        rotationMatrixY[2][2] = (float) Math.cos(angle);
+        rotationMatrixY[0][2] = -Math.sin(angle);
+        rotationMatrixY[1][2] = 0.0;
+        rotationMatrixY[2][2] = Math.cos(angle);
     }
 
-    private void setRotationMatrixXYZ(float angle, final Vector axis) {
-        rotationMatrixXYZ[0][0] = (float) (Math.cos(angle) + (1.0f - Math.cos(angle)) * axis.getX() * axis.getX());
-        rotationMatrixXYZ[1][0] = (float) ((1.0f - Math.cos(angle)) * axis.getY() * axis.getX() - Math.sin(angle) * axis.getZ());
-        rotationMatrixXYZ[2][0] = (float) ((1.0f - Math.cos(angle)) * axis.getZ() * axis.getX() + Math.sin(angle) * axis.getY());
+    private void setRotationMatrixXYZ(double angle, final Vector axis) {
+        rotationMatrixXYZ[0][0] = Math.cos(angle) + (1.0 - Math.cos(angle)) * axis.getX() * axis.getX();
+        rotationMatrixXYZ[1][0] = (1.0 - Math.cos(angle)) * axis.getY() * axis.getX() - Math.sin(angle) * axis.getZ();
+        rotationMatrixXYZ[2][0] = (1.0 - Math.cos(angle)) * axis.getZ() * axis.getX() + Math.sin(angle) * axis.getY();
 
-        rotationMatrixXYZ[0][1] = (float) ((1.0f - Math.cos(angle)) * axis.getX() * axis.getY() + Math.sin(angle) * axis.getZ());
-        rotationMatrixXYZ[1][1] = (float) (Math.cos(angle) + (1.0f - Math.cos(angle)) * axis.getY() * axis.getY());
-        rotationMatrixXYZ[2][1] = (float) ((1.0f - Math.cos(angle)) * axis.getZ() * axis.getY() - Math.sin(angle) * axis.getX());
+        rotationMatrixXYZ[0][1] = (1.0 - Math.cos(angle)) * axis.getX() * axis.getY() + Math.sin(angle) * axis.getZ();
+        rotationMatrixXYZ[1][1] = Math.cos(angle) + (1.0 - Math.cos(angle)) * axis.getY() * axis.getY();
+        rotationMatrixXYZ[2][1] = (1.0 - Math.cos(angle)) * axis.getZ() * axis.getY() - Math.sin(angle) * axis.getX();
 
-        rotationMatrixXYZ[0][2] = (float) ((1.0f - Math.cos(angle)) * axis.getX() * axis.getZ() - Math.sin(angle) * axis.getY());
-        rotationMatrixXYZ[1][2] = (float) ((1.0f - Math.cos(angle)) * axis.getY() * axis.getZ() + Math.sin(angle) * axis.getX());
-        rotationMatrixXYZ[2][2] = (float) (Math.cos(angle) + (1.0f - Math.cos(angle)) * axis.getZ() * axis.getZ());
+        rotationMatrixXYZ[0][2] = (1.0 - Math.cos(angle)) * axis.getX() * axis.getZ() - Math.sin(angle) * axis.getY();
+        rotationMatrixXYZ[1][2] = (1.0 - Math.cos(angle)) * axis.getY() * axis.getZ() + Math.sin(angle) * axis.getX();
+        rotationMatrixXYZ[2][2] = Math.cos(angle) + (1.0 - Math.cos(angle)) * axis.getZ() * axis.getZ();
     }
 
     @Override
-    public void setX(float x) {
+    public void setX(double x) {
         super.setX(x);
         normalize();
     }
 
     @Override
-    public void setY(float y) {
+    public void setY(double y) {
         super.setY(y);
         normalize();
     }
 
     @Override
-    public void setZ(float z) {
+    public void setZ(double z) {
         super.setZ(z);
         normalize();
     }
