+++ src/main/java/com/private_void/core/particles/AtomFactory.java
@@ -3,5 +3,5 @@ package com.private_void.core.particles;
 import com.private_void.core.geometry.CartesianPoint;
 
 public interface AtomFactory {
-    Atom getNewAtom(final CartesianPoint coordinate, float charge);
+    Atom getNewAtom(final CartesianPoint coordinate, double charge);
 }
