+++ src/main/java/com/private_void/core/plates/TorusPlate.java
@@ -4,10 +4,10 @@ import com.private_void.core.geometry.CartesianPoint;
 import com.private_void.core.surfaces.CapillarFactory;
 
 public class TorusPlate extends FlatPlate {
-    private float maxAngleR;
+    private double maxAngleR;
 
-    public TorusPlate(final CapillarFactory capillarFactory, final CartesianPoint center, float capillarsDensity,
-                      float sideLength, float maxAngleR) {
+    public TorusPlate(final CapillarFactory capillarFactory, final CartesianPoint center, double capillarsDensity,
+                      double sideLength, double maxAngleR) {
         super(capillarFactory, center, capillarsDensity, sideLength);
         this.maxAngleR = maxAngleR;
     }
